+++ src/components/inputComponents/Select/SelectOptions.tsx
@@ -1,0 +1,10 @@
+export const Options = () => {
+  return (
+    <>
+      <option value="not selected">Not selected</option>
+      <option value="male">Male</option>
+      <option value="female">Female</option>
+      <option value="other">Other</option>
+    </>
+  );
+};
