+++ src/pages/HookForm.tsx
@@ -4,8 +4,8 @@ import SubmitButton from '../components/SubmitButton';
 import { yupResolver } from '@hookform/resolvers/yup';
 import personSchema from './validation/personSchema';
 import ErrorMessage from '../components/ErrorMessage';
-import { HookGenderSelect } from '../components/inputComponents/GenderSelect';
-import { HookInput } from '../components/inputComponents/BasicInput';
+import HookGenderSelect from '../components/inputComponents/Select/GenderSelect';
+import HookInput from '../components/inputComponents/Input/HookInput';
 import { useAppDispatch } from '../hooks';
 import { toBase64 } from '../models/utils';
 import { updateCards } from '../store/cardsSlice';
