+++ src/pages/StandartForm.tsx
@@ -4,9 +4,9 @@ import { useAppDispatch } from '../hooks';
 import { updateCards } from '../store/cardsSlice';
 import { validateForm } from './validation/validateForm';
 import ErrorMessage from '../components/ErrorMessage';
-import BasicInput from '../components/inputComponents/BasicInput';
+import BasicInput from '../components/inputComponents/Input/BasicInput';
 import SubmitButton from '../components/SubmitButton';
-import GenderSelect from '../components/inputComponents/GenderSelect';
+import GenderSelect from '../components/inputComponents/Select/GenderSelect';
 import { toBase64 } from '../models/utils';
 import { getStrength } from './validation/passwordValidation';
 import Autocomplete from '../components/inputComponents/Autocomplete/Autocomplete';
